feat: complete avatar
